rename: 파일명을 전부  소문자로 교체(#47)
