+++ src/App.tsx
@@ -1,11 +1,11 @@
 import { Route, Routes } from 'react-router-dom'
 import './App.css'
 import MainLayout from '@layouts/MainLayout'
-import TestHub from '@pages/TestPage/TestHub'
-import TestApplicationModalPage from './pages/TestPage/TestApplicationModalPage'
+import TestHub from '@src/pages/test-page/TestHub'
+import TestApplicationModalPage from './pages/test-page/TestApplicationModalPage'
 import { ROUTES } from './constants/routes'
-import RecruitmentPase from '@pages/RecruitmentPage/RecruitmentPage'
-import TestUIPage from './pages/TestPage/TestUIPage'
+import RecruitmentPase from '@src/pages/recruitment-page/RecruitmentPage'
+import TestUIPage from './pages/test-page/TestUIPage'
 
 export default function App() {
   return (
