+++ src/components/chat/ParticipantsList.tsx
